api/auth-app/requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==3.0.2
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
Flask-Login==0.6.3
Werkzeug==2.2.0
markupsafe==2.1.1
WTForms==2.3.3
Flask-WTF==0.15.1
email-validator==1.1.3
